# perf: prepare partial evaluation at boot for better performances