--- a/core/rego.go
+++ b/core/rego.go
@@ -1,0 +1,54 @@
+// Copyright 2024 Mia srl
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package core
+
+import (
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/open-policy-agent/opa/ast"
+	"github.com/open-policy-agent/opa/rego"
+	"github.com/rond-authz/rond/custom_builtins"
+)
+
+func newRegoInstanceBuilder(policy string, opaModuleConfig *OPAModuleConfig, evaluatorOptions *OPAEvaluatorOptions) (*rego.Rego, error) {
+	if opaModuleConfig == nil {
+		return nil, fmt.Errorf("OPAModuleConfig must not be nil")
+	}
+
+	if evaluatorOptions == nil {
+		evaluatorOptions = &OPAEvaluatorOptions{}
+	}
+
+	sanitizedPolicy := strings.Replace(policy, ".", "_", -1)
+	queryString := fmt.Sprintf("data.policies.%s", sanitizedPolicy)
+
+	options := []func(*rego.Rego){
+		rego.Query(queryString),
+		rego.Module(opaModuleConfig.Name, opaModuleConfig.Content),
+		rego.Unknowns(Unknowns),
+		rego.EnablePrintStatements(evaluatorOptions.EnablePrintStatements),
+		rego.PrintHook(NewPrintHook(os.Stdout, policy)),
+		rego.Capabilities(ast.CapabilitiesForThisVersion()),
+		custom_builtins.GetHeaderFunction,
+	}
+	if evaluatorOptions.MongoClient != nil {
+		options = append(options, custom_builtins.MongoFindOne, custom_builtins.MongoFindMany)
+	}
+	regoInstance := rego.New(options...)
+
+	return regoInstance, nil
+}