… core/opaevaluator_test.go
… core/partialevaluators_test.go
… core/rego_test.go