This PR will improve performance of query generation of more than x20, and it seems also the allow and response.

Here the before/after on my machine:

Before: 

```
BenchmarkEvaluateRequest-10                                 4534            255834 ns/op          194671 B/op       3024 allocs/op
BenchmarkEvaluateRequestWithQueryGeneration-10               154           7485473 ns/op         4507957 B/op     112217 allocs/op
BenchmarkEvaluateResponse-10                                1639            702470 ns/op          486613 B/op       7980 allocs/op
```

After:

```
BenchmarkEvaluateRequest-10                                 5054            208840 ns/op          165831 B/op       2437 allocs/op
BenchmarkEvaluateRequestWithQueryGeneration-10              3505            339451 ns/op          241816 B/op       4309 allocs/op
BenchmarkEvaluateResponse-10                                2014            576589 ns/op          418976 B/op       6992 allocs/op
```